+++ moyun-server/src/main/java/com/moyun/ledger/domain/entity/LedgerNetWorthSnapshot.java
@@ -5,7 +5,9 @@ import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
+
 import java.time.LocalDateTime;
 
 /**
@@ -26,14 +28,14 @@ public class LedgerNetWorthSnapshot {
     /** 快照日期（每日定时任务生成；当日有记账实时upsert） */
     private LocalDate snapDate;
 
-    /** 总资产（分） */
-    private Long totalAsset;
+    /** 总资产（元） */
+    private BigDecimal totalAsset;
 
-    /** 总负债（分） */
-    private Long totalLiability;
+    /** 总负债（元） */
+    private BigDecimal totalLiability;
 
-    /** 净资产（分）= total_asset - total_liability */
-    private Long netWorth;
+    /** 净资产（元）= total_asset - total_liability */
+    private BigDecimal netWorth;
 
     private LocalDateTime createTime;
 }
